+++ TerrainGenerator/Models/HydraulicErosion.cs
@@ -10,21 +10,21 @@ namespace Topographer3D.Models
     {
         private int seed;
         private int erosionRadius = 2;
-        private double inertia = 0.025f; // At zero, water will instantly change direction to flow downhill. At 1, water will never change direction. 
-        private double sedimentCapacityFactor = 2; // Multiplier for how much sediment a droplet can carry
-        private double minSedimentCapacity = 0.005; // Used to prevent carry capacity getting too close to zero on flatter terrain
-        private double erodeSpeed = 0.15f;
-        private double depositSpeed = 0.15f;
-        private double evaporateSpeed = 0.005f;
-        private double gravity = 2;
+        private float inertia = 0.025f; // At zero, water will instantly change direction to flow downhill. At 1, water will never change direction. 
+        private float sedimentCapacityFactor = 2; // Multiplier for how much sediment a droplet can carry
+        private float minSedimentCapacity = 0.005f; // Used to prevent carry capacity getting too close to zero on flatter terrain
+        private float erodeSpeed = 0.15f;
+        private float depositSpeed = 0.15f;
+        private float evaporateSpeed = 0.005f;
+        private float gravity = 2;
         private int maxDropletLifetime = 15;
 
-        private double initialWaterVolume = 1;
-        private double initialSpeed = 1;
+        private float initialWaterVolume = 1;
+        private float initialSpeed = 1;
 
         // Indices and weights of erosion brush precomputed for every node
         private int[][] erosionBrushIndices;
-        private double[][] erosionBrushWeights;
+        private float[][] erosionBrushWeights;
         private Random prng;
 
         private int currentSeed;
@@ -48,7 +48,7 @@ namespace Topographer3D.Models
             }
         }
 
-        public void UpdateValues(int seed, int erosionRadius, double inertia, double sedimentCapacityFactor, double minSedimentCapacity, double erodeSpeed, double depositSpeed, double evaporateSpeed, double gravity, int maxDropletLifetime)
+        public void UpdateValues(int seed, int erosionRadius, float inertia, float sedimentCapacityFactor, float minSedimentCapacity, float erodeSpeed, float depositSpeed, float evaporateSpeed, float gravity, int maxDropletLifetime)
         {
             this.seed = seed;
             this.erosionRadius = erosionRadius;
@@ -62,20 +62,20 @@ namespace Topographer3D.Models
             this.maxDropletLifetime = maxDropletLifetime;
         }
 
-        public void Erode(double[] map, int mapSize, int numIterations = 1, bool resetSeed = false)
+        public void Erode(float[] map, int mapSize, int numIterations = 1, bool resetSeed = false)
         {
             Initialize(mapSize, resetSeed);
 
             for (int iteration = 0; iteration < numIterations; iteration++)
             {
                 // Create water droplet at random point on map
-                double posX = prng.Next(0, mapSize - 1);
-                double posZ = prng.Next(0, mapSize - 1);
-                double dirX = 0;
-                double dirY = 0;
-                double speed = initialSpeed;
-                double water = initialWaterVolume;
-                double sediment = 0;
+                float posX = prng.Next(0, mapSize - 1);
+                float posZ = prng.Next(0, mapSize - 1);
+                float dirX = 0;
+                float dirY = 0;
+                float speed = initialSpeed;
+                float water = initialWaterVolume;
+                float sediment = 0;
 
                 for (int lifetime = 0; lifetime < maxDropletLifetime; lifetime++)
                 {
@@ -83,8 +83,8 @@ namespace Topographer3D.Models
                     int nodeY = Convert.ToInt32(posZ);
                     int dropletIndex = nodeY * mapSize + nodeX;
                     // Calculate droplet's offset inside the cell (0,0) = at NW node, (1,1) = at SE node
-                    double cellOffsetX = posX - nodeX;
-                    double cellOffsetY = posZ - nodeY;
+                    float cellOffsetX = posX - nodeX;
+                    float cellOffsetY = posZ - nodeY;
 
 
                     // Calculate droplet's height and direction of flow with bilinear interpolation of surrounding heights
@@ -96,7 +96,7 @@ namespace Topographer3D.Models
                     dirY = (dirY * inertia - heightAndGradient.gradientY * (1 - inertia));
 
                     // Normalize direction
-                    double len = Math.Sqrt(dirX * dirX + dirY * dirY);
+                    float len = (float)Math.Sqrt(dirX * dirX + dirY * dirY);
                     if (len != 0)
                     {
                         dirX /= len;
@@ -113,17 +113,17 @@ namespace Topographer3D.Models
                     }
 
                     // Find the droplet's new height and calculate the deltaHeight
-                    double newHeight = CalculateHeightAndGradient(map, mapSize, posX, posZ).height;
-                    double deltaHeight = newHeight - heightAndGradient.height;
+                    float newHeight = CalculateHeightAndGradient(map, mapSize, posX, posZ).height;
+                    float deltaHeight = newHeight - heightAndGradient.height;
 
                     // Calculate the droplet's sediment capacity (higher when moving fast down a slope and contains lots of water)
-                    double sedimentCapacity = Math.Max(-deltaHeight * speed * water * sedimentCapacityFactor, minSedimentCapacity);
+                    float sedimentCapacity = Math.Max(-deltaHeight * speed * water * sedimentCapacityFactor, minSedimentCapacity);
 
                     // If carrying more sediment than capacity, or if flowing uphill:
                     if (sediment > sedimentCapacity || deltaHeight > 0)
                     {
                         // If moving uphill (deltaHeight > 0) try fill up to the current height, otherwise deposit a fraction of the excess sediment
-                        double amountToDeposit = (deltaHeight > 0) ? Math.Min(deltaHeight, sediment) : (sediment - sedimentCapacity) * depositSpeed;
+                        float amountToDeposit = (deltaHeight > 0) ? Math.Min(deltaHeight, sediment) : (sediment - sedimentCapacity) * depositSpeed;
                         sediment -= amountToDeposit;
 
                         // Add the sediment to the four nodes of the current cell using bilinear interpolation
@@ -140,14 +140,14 @@ namespace Topographer3D.Models
                     {
                         // Erode a fraction of the droplet's current carry capacity.
                         // Clamp the erosion to the change in height so that it doesn't dig a hole in the terrain behind the droplet
-                        double amountToErode = Math.Min((sedimentCapacity - sediment) * erodeSpeed, -deltaHeight);
+                        float amountToErode = Math.Min((sedimentCapacity - sediment) * erodeSpeed, -deltaHeight);
   
                         // Use erosion brush to erode from all nodes inside the droplet's erosion radius
                         for (int brushPointIndex = 0; brushPointIndex < erosionBrushIndices[dropletIndex].Length; brushPointIndex++)
                         {
                             int nodeIndex = erosionBrushIndices[dropletIndex][brushPointIndex];
-                            double weighedErodeAmount = amountToErode * erosionBrushWeights[dropletIndex][brushPointIndex];
-                            double deltaSediment = (map[nodeIndex] < weighedErodeAmount) ? map[nodeIndex] : weighedErodeAmount;
+                            float weighedErodeAmount = amountToErode * erosionBrushWeights[dropletIndex][brushPointIndex];
+                            float deltaSediment = (map[nodeIndex] < weighedErodeAmount) ? map[nodeIndex] : weighedErodeAmount;
                             map[nodeIndex] -= deltaSediment;
                             sediment += deltaSediment;
                         }
@@ -160,7 +160,7 @@ namespace Topographer3D.Models
                     }
                     else
                     {
-                        speed = Math.Sqrt(speed * speed + deltaHeight * gravity);
+                        speed = (float)Math.Sqrt(speed * speed + deltaHeight * gravity);
                     }
 
                     water *= (1 - evaporateSpeed);
@@ -168,30 +168,30 @@ namespace Topographer3D.Models
             }
         }
 
-        private HeightAndGradient CalculateHeightAndGradient(double[] nodes, int mapSize, double posX, double posZ)
+        private HeightAndGradient CalculateHeightAndGradient(float[] nodes, int mapSize, float posX, float posZ)
         {
             int coordX = (int)posX;
             int coordZ = (int)posZ;
 
             // Calculate droplet's offset inside the cell (0,0) = at NW node, (1,1) = at SE node
-            double x = posX - coordX;
-            double y = posZ - coordZ;
+            float x = posX - coordX;
+            float y = posZ - coordZ;
 
             // Calculate heights of the four nodes of the droplet's cell
             int nodeIndexNW = coordZ * mapSize + coordX;
-            double heightNE = nodes[nodeIndexNW + 1];
-            double heightSW = nodes[nodeIndexNW + mapSize];
-            double heightNW = nodes[nodeIndexNW];
-            double heightSE = nodes[nodeIndexNW + mapSize + 1];
+            float heightNE = nodes[nodeIndexNW + 1];
+            float heightSW = nodes[nodeIndexNW + mapSize];
+            float heightNW = nodes[nodeIndexNW];
+            float heightSE = nodes[nodeIndexNW + mapSize + 1];
 
             // Calculate droplet's direction of flow with bilinear interpolation of height difference along the edges
 
-            double gradientX = (heightNE - heightNW) * (1 - y) + (heightSE - heightSW) * y;
-            double gradientY = (heightSW - heightNW) * (1 - x) + (heightSE - heightNE) * x;
+            float gradientX = (heightNE - heightNW) * (1 - y) + (heightSE - heightSW) * y;
+            float gradientY = (heightSW - heightNW) * (1 - x) + (heightSE - heightNE) * x;
 
 
             // Calculate height with bilinear interpolation of the heights of the nodes of the cell
-            double height = heightNW * (1 - x) * (1 - y) + heightNE * x * (1 - y) + heightSW * (1 - x) * y + heightSE * x * y;
+            float height = heightNW * (1 - x) * (1 - y) + heightNE * x * (1 - y) + heightSW * (1 - x) * y + heightSE * x * y;
 
 
             return new HeightAndGradient() { height = height, gradientX = gradientX, gradientY = gradientY };
@@ -200,12 +200,12 @@ namespace Topographer3D.Models
         private void InitializeBrushIndices(int mapSize, int radius)
         {
             erosionBrushIndices = new int[mapSize * mapSize][];
-            erosionBrushWeights = new double[mapSize * mapSize][];
+            erosionBrushWeights = new float[mapSize * mapSize][];
 
             int[] xOffsets = new int[radius * radius * 4];
             int[] yOffsets = new int[radius * radius * 4];
-            double[] weights = new double[radius * radius * 4];
-            double weightSum = 0;
+            float[] weights = new float[radius * radius * 4];
+            float weightSum = 0;
             int addIndex = 0;
 
             for (int i = 0; i < erosionBrushIndices.GetLength(0); i++)
@@ -221,7 +221,7 @@ namespace Topographer3D.Models
                     {
                         for (int x = -radius; x <= radius; x++)
                         {
-                            double sqrDst = x * x + y * y;
+                            float sqrDst = x * x + y * y;
                             if (sqrDst < radius * radius)
                             {
                                 int coordX = centreX + x;
@@ -229,7 +229,7 @@ namespace Topographer3D.Models
 
                                 if (coordX >= 0 && coordX < mapSize && coordY >= 0 && coordY < mapSize)
                                 {
-                                    double weight = 1 - Math.Sqrt(sqrDst) / radius;
+                                    float weight = 1.0f - (float)Math.Sqrt(sqrDst) / radius;
                                     weightSum += weight;
                                     weights[addIndex] = weight;
                                     xOffsets[addIndex] = x;
@@ -243,7 +243,7 @@ namespace Topographer3D.Models
 
                 int numEntries = addIndex;
                 erosionBrushIndices[i] = new int[numEntries];
-                erosionBrushWeights[i] = new double[numEntries];
+                erosionBrushWeights[i] = new float[numEntries];
 
                 for (int j = 0; j < numEntries; j++)
                 {
@@ -255,9 +255,9 @@ namespace Topographer3D.Models
 
         struct HeightAndGradient
         {
-            public double height;
-            public double gradientX;
-            public double gradientY;
+            public float height;
+            public float gradientX;
+            public float gradientY;
         }
     }
 }
